backend/core/config.py
import os
from dotenv import load_dotenv

from pathlib import Path
env_path = Path('.') / '.env'
load_dotenv(dotenv_path=env_path)

class Settings:
    PROJECT_NAME:str = "Lov Lang"
    PROJECT_VERSION: str = "1.0.0"

    POSTGRES_USER : str = os.getenv("POSTGRES_USER")
    POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
    POSTGRES_SERVER : str = os.getenv("POSTGRES_SERVER")
    #POSTGRES_PORT : int = os.getenv("POSTGRES_PORT") # default postgres port is 5432
    POSTGRES_DB : str = os.getenv("POSTGRES_DB")
    DATABASE_URL = "postgresql://postgres:Fesa_1791@127.0.0.1:5433/project"
    
    SECRET_KEY :str = os.getenv("SECRET_KEY")
    #SECRET_KEY =    
    ALGORITHM = "HS256"                         
    ACCESS_TOKEN_EXPIRE_MINUTES = 30  #in mins  

settings = Settings()